+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Update: Renamed "Version" to "ComparableVersion" to avoid name conflicts in including projects.
 # Add ability to specify separator character as an initialization option.
+* Add ComparableVersion#inspect for more informative inspect output.